Improve documentation Move FaceViewDelegate conformance to VatomView
